Added custom hook for worker with avgColors, moved to a hooks folder
